% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region-reorg-fcns.R
\name{abv.rcols}
\alias{abv.rcols}
\title{These functions just manipulate dataframes that contain multiple
regions/regiontypes (i.e., CZs and CBSAs). Some may have general broad use, others
may just help me shift quickly between conventions that I've developed for this
sort of data.}
\usage{
abv.rcols(x)
}
\arguments{
\item{x}{df with columns with name to possibly abbreviate}
}
\description{
abv.rcols
}
\details{
Shortens identifier column names.
 region.name -> rn

 region.id -> rid

 region.type -> rt
}
