% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region-reorg-fcns.R
\name{add.rns}
\alias{add.rns}
\title{add.rns}
\usage{
add.rns(x, abvcols = T)
}
\arguments{
\item{x}{df to add region names to}

\item{abvcols}{whether to abbreviate region id columns}
}
\description{
Add region names to a data.frame long by region.type/region.id (and abbreviates
colnames by default).
}
