% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cbg2plc}
\alias{cbg2plc}
\title{cbg2plc}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 69218 rows and 3 columns.
}
\usage{
cbg2plc
}
\description{
2019 Places with >= 100,000 population, joined with 2019 census block groups
}
\keyword{datasets}
