% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ct_cbsa_xwalk}
\alias{ct_cbsa_xwalk}
\title{ct_cbsa_xwalk}
\format{
An object of class \code{data.frame} with 68199 rows and 3 columns.
}
\usage{
ct_cbsa_xwalk
}
\description{
Unique row/ct with xwalk to cbsas **i should double-check this one
}
\keyword{datasets}
