% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cts2counties_time.series}
\alias{cts2counties_time.series}
\title{cts2counties_time.series}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 79007 rows and 4 columns.
}
\usage{
cts2counties_time.series
}
\description{
Correspondence from census tracts to counties at different years
}
\keyword{datasets}
