% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cts2cts_time.series}
\alias{cts2cts_time.series}
\title{cts2counties_time.series}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 73642 rows and 4 columns.
}
\usage{
cts2cts_time.series
}
\description{
Correspondence from census tracts to counties at different years. Important note:
Two census tracts were split from 2010-2015 (in Alaska). The tract geoid for these
for 2015 are: 02195000200 & 02198000100. If merging, this will cause duplicates
for information pertaining to these two tracts.
}
\details{
From Census Bureau, The part of 02105000300 that became 02195000200 has estimated
detached population of 1; and the part of 02195000200 that became 02198000100 has
estimated added population of 613.

https://www.census.gov/programs-surveys/geography/technical-documentation/county-changes.2010.html
}
\keyword{datasets}
