% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ctx}
\alias{ctx}
\title{ctx}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 74133 rows and 7 columns.
}
\usage{
ctx
}
\description{
Xwalk from 2015 census tracts to 2015 counties, 2019 cbsas, and 1990 czs
}
\keyword{datasets}
