% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{czip}
\alias{czip}
\title{czip}
\format{
An object of class \code{data.frame} with 35980 rows and 4 columns.
}
\usage{
czip
}
\description{
Xwalk between ZIPs and CZs. They're not co-terminous so % overlap is
included. For ZIPs that span multiple CZs, this xwalk will have multiple
rows/ZIP
}
\keyword{datasets}
