% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/region-reorg-fcns.R
\name{fix.geoid}
\alias{fix.geoid}
\title{fix.geoid}
\usage{
fix.geoid(x, width = NULL)
}
\arguments{
\item{x}{a vector of identifiers that can be coerced to character}

\item{width}{If NULL, pads until all elements have width equal to the maximum
width in vector. Otherwise, a numeric specifying width}
}
\description{
Adds leading 0s if necessary and ensures that a given vector is a character
vector. Fixes geoids of any type that may have been read as numeric and turns into
valid identifiers.
}
