% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xwalk-fcns.R
\name{generate.coterminous.xwalk}
\alias{generate.coterminous.xwalk}
\title{generate.coterminous.xwalk}
\usage{
generate.coterminous.xwalk(smaller.geo, larger.geo, trim.smaller = F)
}
\arguments{
\item{smaller.geo}{sf object containing the smaller geometries. (Because this
is for coterminous regions, 1 set should be consistently smaller or the
same size as the other)}

\item{larger.geo}{sf object containing larger geos.}

\item{trim.smaller}{Whether or not to trim smaller geos to combined area of
all larger ones. Necessary for example if larger geos are clipped to water lines or
otherwise trimmed, and the smaller ones are not. (In these cases, the smaller
geometry is known to be smaller, but its spatial representation may not
actually be)}
}
\description{
If two regions are known to be co-terminous, this will generate a xwalk
between them. This method is more reliable than other obvious methods, which
can be thrown off by irregular shapes or differing resolutions between the
spatial datasets. Put the smaller region first. Will return 1 row/smaller
geo.
}
