% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-fcns.R
\name{get.spatial.overlap}
\alias{get.spatial.overlap}
\title{Generate xwalk}
\usage{
get.spatial.overlap(
  shp1,
  shp2,
  shp1.title = "GEOID10",
  shp2.title = "CBSAFP",
  filter.threshold = 0.4
)
}
\description{
This function takes intersections and then trims out areas that intersect a
percentage area below \code{filter.threshold}. Generates a xwalk given two
shapefiles and colnames for unique identifiers for respective shapefiles.
Returns a table with the percent intersection of each area. Filters out
boundary cases where the intersection is a fraction of a percent of total
area This filter threshold takes out edge cases (when they touch but don't
overlap or where low res shapefile would lead to mis-read). .4 may seem high,
but there a couple jigsaw CBSAs in the country that (with a low-res
shapefile) cause duplicate cbsa-ct combos if the filter threshold is lower.
--See Boulder+Denver as example
}
