% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plc.co.cz}
\alias{plc.co.cz}
\title{plc.co.cz}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2156 rows and 5 columns.
}
\usage{
plc.co.cz
}
\description{
Correspondences between 2019 Places where population >= 100,000, with 1990 CZs, and 2015 counties. Percentage
overlap is included because in rare cases, Places span counties and/or CZs.
}
\keyword{datasets}
