% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geosubset-fcns.R
\name{plc2co}
\alias{plc2co}
\title{plc2co}
\usage{
plc2co(plc_id, overlap.threshold = 50)
}
\arguments{
\item{plc_id}{Identifier for region to get water for.}

\item{overlap.threshold}{If a CBG has >= this % inside a Place boundaries, it will
be included.}
}
\description{
Helper fcn called from `x2cos` when a plc_id is provided
}
