% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{state2div}
\alias{state2div}
\title{state2div}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 51 rows and 7 columns.
}
\usage{
state2div
}
\description{
Unique row/state w/ xwalk to larger areas: divisions, regions. Census Bureau
definition:
https://www2.census.gov/geo/pdfs/maps-data/maps/reference/us_regdiv.pdf
}
\keyword{datasets}
