% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geosubset-fcns.R
\name{x2cos}
\alias{x2cos}
\title{x2cos}
\usage{
x2cos(
  cz_name = NULL,
  cz = NULL,
  countyfp = NULL,
  cbsa_id = NULL,
  plc_id = NULL
)
}
\arguments{
\item{cz_name, cz, countyfp, cbsa_id, plc_id}{Identifier for region to get water for.}
}
\value{
List of 5-char county fp codes overlapping with region.
}
\description{
Given a cz,cbsa, or place identifier/geoid, gets all counties (5-character fp codes)
that overlap given region. countyfps correspond with 2015 counties.
}
