% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{co2cz}
\alias{co2cz}
\title{co2cz}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3142 rows and 4 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3142 rows and 4 columns.
}
\usage{
co2cz

co2cz
}
\description{
2015 counties w/ unique row/county with xwalk to state and 1990 cz.

2015 counties w/ unique row/county with xwalk to 2019 cbsa and state.
}
\keyword{datasets}
