% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xwalk-fcns.R
\name{get.spatial.overlap}
\alias{get.spatial.overlap}
\title{get.spatial.overlap}
\usage{
get.spatial.overlap(
  sf1,
  sf2,
  sf1.identifier,
  sf2.identifier,
  filter.threshold = 0.01
)
}
\arguments{
\item{sf1}{First sf object}

\item{sf2}{Second sf object}

\item{sf1.identifier}{colname as string for region identifiers for first sf
object}

\item{sf2.identifier}{colname as string for region identifiers for second sf
object}

\item{filter.threshold}{percent overlap between sf1 and sf2, as decimal,
below which to trim results before returning.}
}
\description{
Calculates the % overlap between two geographies. Takes two sf objects, a
unique/row identifier column for each; returns a row for each combination of
rows in the input sfs that intersect with one another with a column for the
percent overlap. % overlap can indicate true overlap, or it can be an
artifact of differing resolutions between the shapefiles or other artifacts
of their representation. Filtering to geometries below a very small % overlap
using \code{filter.threshold} or otherwise can remove these slivers.
}
