% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xwalk-fcns.R
\name{get.spatial.overlap}
\alias{get.spatial.overlap}
\title{get.spatial.overlap}
\usage{
get.spatial.overlap(
  sf1,
  sf2,
  sf1.identifier,
  sf2.identifier,
  filter.threshold = 0.01
)
}
\arguments{
\item{sf1}{First sf object}

\item{sf2}{Second sf object}

\item{sf1.identifier}{colname as string for region identifiers for first sf
object}

\item{sf2.identifier}{colname as string for region identifiers for second sf
object}

\item{filter.threshold}{percent overlap between sf1 and sf2, as decimal, below which to trim
results before returning.}
}
\description{
This takes intersections btwn two sf objects and filters out areas
combinations where the intersect area is below \code{filter.threshold}
percent of the first sf. Returns a table with the percent intersection of
each area. Filters out boundary cases where the intersection is a fraction of
a percent of total area. Filter threshold takes out edge cases (when they
touch but don't overlap or where low res shapefile would lead to mis-read).
By default this is very low.
}
